Move the index builder into a new ann package
Vector-index code is growing beyond one file: the builder is being
joined by a sidecar reader (and later an index evaluator), none of which
is about producing records or embeddings. Give it a package:
indexing/ann.py becomes ann/build.py, with the reader and evaluator to
follow as siblings. Mechanical move; no behavior change.
