+++ src/okn_embeddings/ann/__init__.py
@@ -1,0 +1,2 @@
+"""ANN sidecar artifacts: building, reading, and (eventually) evaluating
+usearch indexes over embed Parquet files."""
